Apply Dep Inversion In Transaction class
What: Created an Operation interface with the
main public methods of the Transaction class
and applied
- Created a new enum for the operation types
Why: So the High-Level class Transactions
Processor doesn't directly depend on the low
level class Transaction, but depend on an
interface that specifies the methods expected
by the specific implementation of Operation.
- Since the Transaction class no longer should
be used in the TransactionProcessor, we didn't
want to keep relying on the static definitions
of BUY and SELL from that class.